backend/foodgram/requirements.txt
django==4.1.2
django-colorfield==0.7.2
djangorestframework==3.12.4
